rpc, smartcontract: move contract metadata to smartcontract package
